requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
wordcloud==1.9.2
nltk==3.8.1
streamlit==1.32.0
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.0
spacy==3.7.2
groq==0.4.2
streamlit-extras==0.3.6